LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/highlight-currentcolor-computed-visited-expected.txt


PASS getComputedStyle() for ::selection at #target1
PASS getComputedStyle() for ::selection at #target2
FAIL getComputedStyle() for ::target-text at #target1 assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got ""
FAIL getComputedStyle() for ::target-text at #target2 assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got ""
PASS getComputedStyle() for ::spelling-error at #target1
PASS getComputedStyle() for ::spelling-error at #target2
PASS getComputedStyle() for ::grammar-error at #target1
PASS getComputedStyle() for ::grammar-error at #target2
FAIL getComputedStyle() for ::highlight(foo) at #target1 assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got ""
FAIL getComputedStyle() for ::highlight(foo) at #target2 assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got ""
